pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-oraclefusion"
version = "0.1.0"
description = "Oracle Fusion Cloud ERP — read-only MCP server (STDIO transport)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.24",
    "python-dotenv",
    "python-jose[cryptography]",
    "cryptography",
]

[project.scripts]
mcp-oraclefusion = "stdio_server:main"

[project.urls]
Repository = "https://github.com/elabbarw/mcp-oraclefusion"

[tool.hatch.build.targets.wheel]
packages = ["oracle_fusion_mcp"]
include = ["/stdio_server.py"]
